Setup SASS and make folders
