+++ .vscode/extensions.json
@@ -1,0 +1,9 @@
+{
+  "recommendations": [
+    "dbaeumer.vscode-eslint",
+    "wesbos.theme-cobalt2",
+    "formulahendry.auto-rename-tag",
+    "graphql.vscode-graphql",
+    "jpoissonnier.vscode-styled-components"
+  ]
+}
