+++ src/App/App.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Navbar from './components/Navbar';
-import Home from './components/pages/Home';
+import './App.scss';
+import Navbar from '../components/NavBar/Navbar';
+import Home from '../pages/Home';
 
 function App() {
   return (
