+++ src/App/App.scss
@@ -1,0 +1,1 @@
+@import '../../node_modules/modern-normalize/modern-normalize.css';
