+++ src/components/NavBar/Navbar.scss
@@ -1,0 +1,9 @@
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  background-color: papayawhip;
+  height: 100px;
+  padding: 10px;
+}
